Move ConnectionService tests into a single file
